More error checks

+++ source/philo/src/philosophers/philosophers_utils.h
@@ -6,7 +6,7 @@
 /*   By: yademirk <yademirk@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/23 20:46:31 by yademirk          #+#    #+#             */
-/*   Updated: 2026/03/14 21:56:47 by yademirk         ###   ########.fr       */
+/*   Updated: 2026/03/15 22:50:33 by yademirk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ int		should_philo_continue(t_philosopher *philo);
  * @return 1 on success, 0 on failure.
  */
 int		init_philosophers(t_philosopher *philos, t_table *table,
-			int philo_count);
+			size_t philo_count);
 
 /**
  * @brief Calls pthread_join on all philosophers.
